Split monolithic procmon-mcp.py into procmon_mcp package
Refactor the 2,354-line single-file server into 12 focused modules:
- constants.py: all constants and interner keys
- compat.py: feature detection (lxml, psutil, MCP SDK) with mocks
- helpers.py: XML helpers, regex, timestamp parsing, byte formatting
- models.py: StringInterner, StackFrame, ProcessInfo, ProcmonLogData
- parser.py: two-pass XML parsing and load_procmon_xml
- filters.py: async event filtering engine
- formatters.py: event detail formatting
- server.py: MCP server instance, global state, _check_loaded
- tools.py: all 14 MCP tool definitions
- cli.py: argument parsing and main_execution entry point
- __init__.py: package re-exports for backward compatibility
- __main__.py: python -m procmon_mcp entry point

The original procmon-mcp.py is now a thin wrapper that delegates
to procmon_mcp.cli.main(), preserving backward compatibility for
existing MCP configurations that reference procmon-mcp.py directly.

Tests updated to import from procmon_mcp package. All 68 tests pass.
