+++ procmon_mcp/__main__.py
@@ -1,0 +1,4 @@
+"""Entry point for python -m procmon_mcp."""
+from .cli import main
+
+main()
